one working implementation of temp scaling but needs to be improved
